created a way to iterate over multiple layers, first MultiLayerReader will read the file, create the projection object that can be used to transform each layer. LayerTransformer will then take one layer, and do the projection for us. Should allow us to do all ingestion one layer at a time and be a bit more mem efficient.
a few changes to keep the root project from publishing both locally and otherwise, from creating artifacts that people may confuse with working jars, and the jar it does create for root during assembly is hidden and does not contain .jar extension so as not to confuse anyone into running it.

removed style checking on tests, makes it hard to read what tests are succeeding and/or where failing, updated data to main root project level as some tests seems to find files there. I would like to add that we should be testing and getting data from resource. I will work on that in a later date.

reset fork in tests to be true.

set method back to private.

added deprecation remark

style check error solved

added a test for the iterator, made small adjustment to how the iteration process works to make things cleaner.